Feat: Implement deploying, updating, deleting

+++ app/services/snippet.ts
@@ -1,5 +1,6 @@
 import { createBrowserClient } from "@supabase/ssr";
 
+const snippetLogEndPoint = process.env.NEXT_PUBLIC_LOG_ENDPOINT;
 const supabase = createBrowserClient(
   process.env.NEXT_PUBLIC_SUPABASE_URL!,
   process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY!
@@ -20,15 +21,31 @@ export async function fetchSnippetById(id: string) {
   const { data, error } = await supabase
     .from("snippet")
     .select("*")
-    .eq("id", Number(id));
+    .eq("id", id)
+    .single();
 
   if (error) throw error;
 
   return data;
 }
 
-export async function insertSnippet(props: SnippetDetails) {
+export async function deployAndInsertSnippet(props: SnippetDetails) {
   const { name, code, userId } = props;
+
+  const deployRes = await fetch(`${snippetLogEndPoint}/snippet/deploy`, {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify({
+      function_name: name,
+      script: code,
+    }),
+  });
+
+  if (!deployRes.ok) {
+    throw new Error("Failed to deploy snippet");
+  }
 
   const { data, error } = await supabase
     .from("snippet")
@@ -36,11 +53,26 @@ export async function insertSnippet(props: SnippetDetails) {
 
   if (error) throw error;
 
-  return data;
+  return { deployData: await deployRes.json(), insertData: data };
 }
 
-export async function updateSnippet(props: SnippetDetails) {
+export async function updatedSnippet(props: SnippetDetails) {
   const { id, name, code } = props;
+
+  const updateRes = await fetch(`${snippetLogEndPoint}/snippet/update`, {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify({
+      function_name: name,
+      script: code,
+    }),
+  });
+
+  if (!updateRes.ok) {
+    throw new Error("Failed to update snippet");
+  }
 
   const { data, error } = await supabase
     .from("snippet")
@@ -50,10 +82,26 @@ export async function updateSnippet(props: SnippetDetails) {
 
   if (error) throw error;
 
-  return data;
+  return { updateData: await updateRes.json(), supabaseUpdateData: data };
 }
 
-export async function deleteSnippet(id: string) {
+export async function deletedSnippet(props: SnippetDetails) {
+  const { id, name } = props;
+
+  const deleteRes = await fetch(`${snippetLogEndPoint}/snippet/delete`, {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify({
+      function_name: name,
+    }),
+  });
+
+  if (!deleteRes.ok) {
+    throw new Error("Failed to delete snippet");
+  }
+
   const { data, error } = await supabase
     .from("snippet")
     .delete()
@@ -61,5 +109,21 @@ export async function deleteSnippet(id: string) {
 
   if (error) throw error;
 
-  return data;
+  return { deleteData: await deleteRes.json(), supabaseDeleteData: data };
+}
+
+export async function getLog(functionName: string) {
+  const res = await fetch(`${snippetLogEndPoint}/log`, {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify({ function_name: functionName }),
+  });
+
+  if (!res.ok) {
+    throw new Error("Failed to fetch data");
+  }
+
+  return res.json();
 }
